make tree model semi polymorph and move files to appropriate folder
