+++ src/Structures/TreeModel.php
@@ -4,11 +4,11 @@ namespace Statamic\Eloquent\Structures;
 
 use Illuminate\Database\Eloquent\Model as Eloquent;
 
-class NavTreeModel extends Eloquent
+class TreeModel extends Eloquent
 {
     protected $guarded = [];
 
-    protected $table = 'navigation_trees';
+    protected $table = 'trees';
 
     protected $casts = [
         'tree' => 'json',
